Add files via upload
This code is the part of the YouTube Video.
https://youtu.be/ssxE9PHA8NI
In this video we have used the tinkercad online simulator to perform the interfacing and 
consequently the serial communication between Arduino UNO and ESP8266-01 WiFi module. Two analog sensors TMP36 and Photo-resistor (LDR) are also connected to Arduino UNO. IoT Web Server ThingSpeak.com is configured for visualization of these two sensor data. This video can help for learning of IoT implementation using Arduino and ESP8266 WiFi module along with ThingSpeak web-server data visualization.